+++ changelog.md
@@ -4,6 +4,7 @@ GDB+ 7.0.1: catchpoints
 * allow different register calling conventions for call() 
 * try to trace function calls
 * breakpoint callbacks are executed in LIFO order
+* gdbserver starts from entry point by default
 
 GDB+ 7.0.0: Fast GDB
 * increased speed up to 10x